test(integration): de-duplicate get_db_max_written_values into shared helpers
The new async-replication-over-TLS test copied the multi-unit max-value helper verbatim from the existing async-replication test. Hoisting one copy into high_availability_helpers_new.py removes the duplication and keeps both tests in sync.
